+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Added account code to `miden account --show` command (#835).
 * Changed exec's input file format to TOML instead of JSON (#870).
 * [BREAKING] Client's methods renamed after `PartialMmr` change to `PartialBlockchain` (#894).
+* [BREAKING] Made the maximum number of blocks the client can be behind the network customizable (#895).
 
 ## 0.8.2 (TBD)